summarizer/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
openai==1.75.0
python-dotenv==1.1.0
